feat: 검색바 컴포넌트 분리후 ref 적용 및 비디오 정보 페이지에서도 사용
