+++ src/components/Loading.jsx
@@ -1,0 +1,3 @@
+export default function Loading() {
+  return <div>Loading</div>;
+}
